+++ app/src/pages/skjema/components/SkjemaVeiledning.tsx
@@ -16,7 +16,7 @@ import { useQuery } from "@tanstack/react-query";
 import { useNavigate } from "@tanstack/react-router";
 import { useTranslation } from "react-i18next";
 
-import { getUserInfo } from "~/api/userInfo.ts";
+import { getUserInfo } from "~/httpClients/dekoratorenClient.ts";
 
 interface SkjemaVeiledningProps {
   startRoute?: string;
